+++ src/main/java/svnserver/server/command/LockCmd.java
@@ -58,7 +58,6 @@ public final class LockCmd extends BaseCmd<LockCmd.Params> {
     final String path = context.getRepositoryPath(args.path);
     final LockTarget lockTarget = new LockTarget(path, rev);
     final String comment = args.comment.length == 0 ? null : args.comment[0];
-    context.checkWrite(path);
     final LockDesc[] lockDescs = context.getRepository().wrapLockWrite((lockManager) -> lockManager.lock(context, comment, args.stealLock, new LockTarget[]{lockTarget}));
     if (lockDescs.length != 1) {
       throw new IllegalStateException();
@@ -72,6 +71,11 @@ public final class LockCmd extends BaseCmd<LockCmd.Params> {
     writer
         .listEnd()
         .listEnd();
+  }
+
+  @Override
+  protected void permissionCheck(@NotNull SessionContext context, @NotNull Params args) throws IOException, SVNException {
+    context.checkWrite(context.getRepositoryPath(args.path));
   }
 
   static void writeLock(@NotNull SvnServerWriter writer, @Nullable LockDesc lockDesc) throws IOException {